+++ source/redux/client/client.js
@@ -1,12 +1,15 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import client_side_render, { always_instrument_history_pop_state_listeners, create_history, authentication_token } from '../../client'
+import client_side_render, { should_instrument_history_pop_state_listeners, create_history, authentication_token } from '../../client'
 import render from './render'
 import { create_http_client } from '../http client'
 import normalize_common_settings from '../normalize'
 import create_store from '../store'
 import { preload_action } from '../actions'
+// import { load_state_action } from '../actions'
+import { get_from_history } from './history store'
+// import { store_in_history } from './history store'
 
 // This function is what's gonna be called from the project's code on the client-side.
 export default function set_up_and_render(settings, options = {})
@@ -25,9 +28,37 @@ export default function set_up_and_render(settings, options = {})
 	let store
 
 	// Intercept `popstate` DOM event to preload pages before showing them
-	always_instrument_history_pop_state_listeners((listener, event, location) =>
+	should_instrument_history_pop_state_listeners((listener, event, location) =>
 	{
-		// Preload the page but don't navigate to it
+		// This idea was discarded because state JSON could be very large.
+		// // Store the current Redux state in history
+		// // before performing the "Back"/"Forward" navigation.
+		// store_in_history('redux/state', get_current_location().key, store.getState())
+		//
+		// const redux_state = get_from_history('redux/state', event.state.key)
+		//
+		// if (redux_state)
+		// {
+		// 	// Won't preload the page again but will instead use
+		// 	// the Redux state that was relevant at the time
+		// 	// the page was navigated from.
+		// 	store.dispatch(load_state_action(redux_state))
+		//
+		// 	// Navigate to the page
+		// 	listener(event)
+		// 	return
+		// }
+
+		const location_key = event.state ? event.state.key : 'initial'
+
+		if (get_from_history('instant-back', location_key) === get_current_location().key)
+		{
+			// Navigate to the page without preloading it
+			// (has been previously preloaded and is in Redux state)
+			return listener(event)
+		}
+
+		// Preload the page but don't navigate to it just yet
 		store.dispatch(preload_action(location, undefined, false)).then((result) =>
 		{
 			// If preload was cancelled, then don't call the wrapped listener
@@ -35,7 +66,8 @@ export default function set_up_and_render(settings, options = {})
 			{
 				return
 			}
-
+		
+			// Navigate to the page
 			listener(event)
 		})
 	})
@@ -57,6 +89,20 @@ export default function set_up_and_render(settings, options = {})
 
 	// Create `react-router` `history`
 	history = create_history(document.location, settings.history, { store })
+
+	// When `popstate` event listener is fired,
+	// `history.getCurrentLocation()` is already
+	// the `pop`ped one (for some unknown reason),
+	// therefore the "previous location" could be obtained
+	// using this listener.
+
+	let current_location = history.getCurrentLocation()
+	const get_current_location = () => current_location
+	
+	history.listen((location) =>
+	{
+		current_location = location
+	})
 
 	// Call `onNavigate` on initial page load
 	if (on_navigate)